Dict implementation : string only hashmap (#12)
* Revise string header
* Merge ctype source files in a single file
* Add 5 standard string hashing functions

+++ test/heap/02_push_multiple_elements_test.c
@@ -7,11 +7,11 @@
 
 int push_multiple_elements_test(void)
 {
-	t_heap max_heap __attribute__((cleanup(heap_free))) = HEAP_NEW(int, &spaceship(descending, int));
-	t_heap min_heap __attribute__((cleanup(heap_free))) = HEAP_NEW(int, &spaceship(ascending, int));
-	int    elements[] = {42, 13, 27, 19, 21, 54, 7, 23, -2, 18, -8, 10};
-	const int length  = C_ARRAY_LENGTH(elements);
-	int       i       = length;
+	t_heap       max_heap __attribute__((cleanup(heap_free))) = HEAP_NEW(int, &spaceship(descending, int));
+	t_heap       min_heap __attribute__((cleanup(heap_free))) = HEAP_NEW(int, &spaceship(ascending, int));
+	int          elements[]                                   = {42, 13, 27, 19, 21, 54, 7, 23, -2, 18, -8, 10};
+	const size_t length                                       = C_ARRAY_LENGTH(elements);
+	int          i                                            = length;
 
 	while (i-- > 0)
 	{
